Refactor redis_route_cache.rs to extract tests and make GenericRedisRouteCache crate-private
Extract the large embedded #[cfg(test)] module from redis_route_cache.rs
into separate files to bring the production file under the 400-line limit:
- Move FakeProvider and assert_put_get_round_trips to test_helpers.rs
- Move mock-based unit tests to tests/mock_tests.rs
- Move live Redis tests to tests/live_tests.rs

Make GenericRedisRouteCache and its provider generic parameter crate-private
(pub(crate)) to prevent leaking the provider type in the public API.
RedisRouteCache<P> remains the only public type.

All existing tests continue to pass (6 mock tests run by default,
3 live tests marked #[ignore] for opt-in execution).
